Rename generated parser to lezer
